pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codexopt"
version = "0.1.1"
description = "CodexOpt: Optimize your Agents.MD and Skills for Codex with GEPA"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Shashi", email = "shashi@super-agentic.ai" }]
dependencies = [
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5.0",
  "build>=1.2.0",
]

[project.scripts]
codexopt = "codexopt.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
